AG-7323 - Rework test canvas mocking to allow use outside Jest tests.
