+++ charts-packages/ag-charts-community/src/util/window.ts
@@ -2,7 +2,11 @@ export function windowValue(name: string): undefined | any {
     /**
      * Redeclaration of window that is safe for use with Gatsby server-side (webpack) compilation.
      */
-    const WINDOW = typeof window === 'undefined' ? undefined : (window as any);
+    const WINDOW =
+        typeof window !== 'undefined'
+            ? (window as any)
+            : // typeof global !== 'undefined' ? (global as any) :
+              undefined;
 
     return WINDOW?.[name];
 }
